Add files via upload
there are two Tasks as a begineer level
Q1 program that asks the user to input the following details of a student 
student_ID, name, email_address, department, phone_number
 Then, save these entered details in a file named “student.txt” and ask the user if he/she 
wants to enter any new record. If the user enters ‘Y’, then repeat the process again and 
save/append the newly entered record’s details in the file. If the user enters ‘N’, then the 
program should exit

Q2  program for billing system of a restaurant. The program should perform the 
following tasks:
• Show menu of items/dishes to customers
It should allow customers to select multiple items (by showing menu again and again 
continuously until he/she selects the “Quit” option)
• Finally, when the customer quits, it should prepare the bill (containing the details of all 
the orders and the total) and save it in a file named “bills.txt”.